+++ NEWS.md
@@ -1,3 +1,4 @@
 # ATdatabase 0.0.1
 
 * Start of development
+* Add `list_doc()` to list all references of meta documents for a given type (#25)